[CSPM] Add support for XCCDF benchmarks
This change adds support for XCCDF benchmarks in the pkg/compliance
package.

XCCDF is an XML format specifying security benchmarks, which is part
of the SCAP standards.

The XCCDF support relies on the OpenSCAP library, which is provided
as part of a previous change.

The XCCDF files should be located into the compliance directory,
specified by the "compliance_config.dir" configuration option.

Support of XCCDF benchmark is disabled by default, and can be
enabled by setting the new "compliance_config.xccdf.enabled"
configuration option.

This change is the result of a collaborative work between
Sylvain Baubeau, Pierre Guilleminot and David du Colombier.

Co-authored-by: Pierre Guilleminot <pierre.guilleminot@datadoghq.com>
Co-authored-by: David du Colombier <djc@datadoghq.com>